% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userinput.R
\name{Confirm}
\alias{Confirm}
\title{Prompts the user for confirmation in the console}
\usage{
Confirm(msg, default = NULL)
}
\arguments{
\item{msg}{string the thing that you are want a yes/no answer to}

\item{default}{mixed optional. If supplied and y or yes or similar, will default to true.
If supplied and not similar to yes will default to no.
If not supplied, will not have a default and the user must choose an option}
}
\value{
boolean
}
\description{
Prompts the user for confirmation in the console
}
\details{
Presents the user with a yes/no question and returns true or false depending on their answer
}

