% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userinput.R
\name{GetMultiUserchoice}
\alias{GetMultiUserchoice}
\title{allows the user to select 1 or more of the choices,}
\usage{
GetMultiUserchoice(options, choosing.what = "one of the following",
  default = 1, all = FALSE)
}
\arguments{
\item{options}{string vector; list of choices}

\item{choosing.what}{string; instrucitons for user}

\item{default}{int or string "all"; which options should be selected if the just hits clicks 'enter'}

\item{all}{boolean; should there be an extra option at the end to choose all the options in the list?}
}
\value{
int vector of the choice numbers
}
\description{
allows the user to select 1 or more of the choices,
}

