% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userinput.R
\name{Preset}
\alias{Preset}
\title{sets the preset input global variable, which if not empty will be used instead
of .ReadLine.}
\usage{
Preset(user.input.strings = character())
}
\arguments{
\item{user.input.strings}{character}
}
\description{
Allows tests to preset the userinput without interrupting the test with .ReadLine.
}
\details{
This should probably not be used except for its designed purpose of unit
tests on scripts that use userinput. It could cause unexpected behaviour if
by mistake something is left in the preset.input variable. Use on.exit(Preset())
}

