% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userinput.R
\name{ReadInt}
\alias{ReadInt}
\title{Reads an int input from the user and re-prompts if they didn't enter an int}
\usage{
ReadInt(msg = "Enter an integer", min = 1, max = NA, default = NULL)
}
\arguments{
\item{msg}{character}

\item{min}{int}

\item{max}{int}

\item{default}{int optional}
}
\description{
Reads an int input from the user and re-prompts if they didn't enter an int
}

